Start app from seperate file to app.ts
